api/requirements.txt
Flask
flask-cors
python-dotenv
python-jose
six
psycopg2